Fix new API functions
Fixed new API functions to be used in the new commands
Fixed the location of the new commands to the native_commands folder
